Change for loops to int from byte becuase of what leagris said

+++ src/main/java/gregtech/api/enums/Textures.java
@@ -1485,8 +1485,8 @@ public class Textures {
         }
 
         static {
-            for (byte i = 0; i < MACHINE_CASINGS.length; i++)
-                for (byte j = 0; j < MACHINE_CASINGS[i].length; j++)
+            for (int i = 0; i < MACHINE_CASINGS.length; i++)
+                for (int j = 0; j < MACHINE_CASINGS[i].length; j++)
                     MACHINE_CASINGS[i][j] = new GT_SidedTexture(MACHINECASINGS_BOTTOM[i], MACHINECASINGS_TOP[i], MACHINECASINGS_SIDE[i], Dyes.getModulation(j - 1, Dyes.MACHINE_METAL.mRGBa));
             casingTexturePages[0] = CASING_BLOCKS;
             //adds some known pages, modders also can do it...
